aug_16_2020
Learned about singletons and reorganized and repaired the game. It's mostly back to basic functionality, except that the loaded scenes with buttons are not clickable. Will work on that tomorrow. Note to self, I checked the PAUSE settings, and nothing is paused (most common similar error on google). So it's a different issue.